fix: address review findings — dedupe haversine, remove dead code, add cache TTL, fix AbortController (023, 024, 025, 026, 029)
