+++ src/types/index.ts
@@ -82,7 +82,7 @@ export interface BlockMetrics {
   resolutionRate: number;
   avgDaysToResolve: number | null;
   topIssues: { category: string; count: number }[];
-  recentlyResolved: { category: string; date: string }[];
+  recentlyResolved?: { category: string; date: string }[];
   radiusMiles: number;
   nearbyOpenIssues?: NearbyOpenIssue[];
   nearbyResources?: NearbyResource[];
